+++ tests/test_view.py
@@ -1,13 +1,16 @@
 # encoding: utf-8
 """ Tests for view utilities """
-import unittest
-from mock import MagicMock, patch, call
+from mock import MagicMock
 from pyramid.httpexceptions import HTTPFound
 from pyramid.testing import DummyRequest
 from pyramid_duh.params import argify
 from pyramid_duh.view import SubpathPredicate, addslash, includeme
-import pyramid_duh
-from pyramid.config import Configurator
+
+
+try:
+    import unittest2 as unittest  # pylint: disable=F0401
+except ImportError:
+    import unittest
 
 
 class TestSubpath(unittest.TestCase):
@@ -25,6 +28,13 @@ class TestSubpath(unittest.TestCase):
         self.request.subpath = ('mypath',)
         result = matcher(None, self.request)
         self.assertTrue(result)
+
+    def test_glob_match_case(self):
+        """ Glob maching respects case """
+        matcher = SubpathPredicate(('foo'), None)
+        self.request.subpath = ('FOO',)
+        result = matcher(None, self.request)
+        self.assertFalse(result)
 
     def test_len_mismatch(self):
         """ Subpath mismatch if incorrect length """